Move code to keptn package
