+++ internal/common/common.go
@@ -13,40 +13,10 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	keptncommon "github.com/keptn/go-utils/pkg/lib"
-	"github.com/keptn/go-utils/pkg/lib/keptn"
 )
 
 const BridgeLabel = "Keptns Bridge"
 const ProblemURLLabel = "Problem URL"
-
-const shipyardControllerURLEnvironmentVariableName = "SHIPYARD_CONTROLLER"
-const configurationServiceEnvironmentVariableName = "CONFIGURATION_SERVICE"
-const datastoreEnvironmentVariableName = "DATASTORE"
-
-const defaultShipyardControllerURL = "http://shipyard-controller:8080"
-
-// GetConfigurationServiceURL Returns the endpoint to the configuration-service
-func GetConfigurationServiceURL() string {
-	return getKeptnServiceURL(configurationServiceEnvironmentVariableName, keptn.ConfigurationServiceURL)
-}
-
-// GetDatastoreURL Returns the endpoint to the datastore
-func GetDatastoreURL() string {
-	return getKeptnServiceURL(datastoreEnvironmentVariableName, keptn.DatastoreURL)
-}
-
-// GetShipyardControllerURL Returns the endpoint to the shipyard-controller
-func GetShipyardControllerURL() string {
-	return getKeptnServiceURL(shipyardControllerURLEnvironmentVariableName, defaultShipyardControllerURL)
-}
-
-func getKeptnServiceURL(serviceName, defaultURL string) string {
-	url, err := keptn.GetServiceEndpoint(serviceName)
-	if err != nil {
-		return defaultURL
-	}
-	return url.String()
-}
 
 // DynatraceConfigDashboardQUERY defines the Dynatrace Configuration File structure and supporting Constants
 const DynatraceConfigDashboardQUERY = "query"
