+++ internal/keptn/client_factory.go
@@ -3,7 +3,6 @@ package keptn
 import (
 	"net/http"
 
-	"github.com/keptn-contrib/dynatrace-service/internal/common"
 	api "github.com/keptn/go-utils/pkg/api/utils"
 )
 
@@ -26,22 +25,22 @@ func NewClientFactory() *ClientFactory {
 
 // CreateEventClient creates an EventClientInterface.
 func (c *ClientFactory) CreateEventClient() EventClientInterface {
-	return NewEventClient(api.NewEventHandler(common.GetDatastoreURL()))
+	return NewEventClient(api.NewEventHandler(GetDatastoreURL()))
 }
 
 // CreateResourceClient creates a ResourceClientInterface.
 func (c *ClientFactory) CreateResourceClient() ResourceClientInterface {
-	return NewResourceClient(api.NewResourceHandler(common.GetConfigurationServiceURL()))
+	return NewResourceClient(api.NewResourceHandler(GetConfigurationServiceURL()))
 }
 
 // CreateServiceClient creates a ServiceClientInterface.
 func (c *ClientFactory) CreateServiceClient() ServiceClientInterface {
 	return NewServiceClient(
-		api.NewServiceHandler(common.GetShipyardControllerURL()),
-		api.NewAuthenticatedAPIHandler(common.GetShipyardControllerURL(), "", "", &http.Client{}, "http"))
+		api.NewServiceHandler(GetShipyardControllerURL()),
+		api.NewAuthenticatedAPIHandler(GetShipyardControllerURL(), "", "", &http.Client{}, "http"))
 }
 
 // CreateUniformClient creates a UniformClientInterface.
 func (c *ClientFactory) CreateUniformClient() UniformClientInterface {
-	return NewUniformClient(api.NewUniformHandler(common.GetShipyardControllerURL()))
+	return NewUniformClient(api.NewUniformHandler(GetShipyardControllerURL()))
 }
