+++ internal/keptn/keptn_client.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/cloudevents/sdk-go/v2/event"
 	"github.com/keptn-contrib/dynatrace-service/internal/adapter"
-	"github.com/keptn-contrib/dynatrace-service/internal/common"
 	keptnapi "github.com/keptn/go-utils/pkg/lib/keptn"
 	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
 )
@@ -87,8 +86,8 @@ func NewClient(client *keptnv2.Keptn) *Client {
 
 func NewDefaultClient(event event.Event) (*Client, error) {
 	keptnOpts := keptnapi.KeptnOpts{
-		ConfigurationServiceURL: common.GetConfigurationServiceURL(),
-		DatastoreURL:            common.GetDatastoreURL(),
+		ConfigurationServiceURL: GetConfigurationServiceURL(),
+		DatastoreURL:            GetDatastoreURL(),
 	}
 	kClient, err := keptnv2.NewKeptn(&event, keptnOpts)
 	if err != nil {
